reorganize front/back end folders and files so it is less confusing
