--- src/Components/backendConnection/AuthProvider.js
@@ -1,15 +1,0 @@
-import { createContext, useState } from "react";
-
-const AuthContext = createContext({});
-
-export const AuthProvider = ({ children }) => {
-    const [auth, setAuth] = useState({});
-
-    return (
-        <AuthContext.Provider value={{ auth, setAuth }}>
-            {children}
-        </AuthContext.Provider>
-    )
-}
-
-export default AuthContext;
