+++ src/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import App from "./App.js";
-import { AuthProvider } from "./Components/backendConnection/AuthProvider";
+import { AuthProvider } from "./Backend/AuthProvider";
 
 ReactDOM.render(
   <React.StrictMode>
